tests: refactor tests to avoid a single huge generate.py
